% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{display_hurwitz_pal}
\alias{display_hurwitz_pal}
\title{Preview color palette for Hurwitz Lab colors}
\usage{
display_hurwitz_pal(name, n = length(hurwitz_palettes[[name]]))
}
\arguments{
\item{name}{Palette name}

\item{n}{Number of colors to display}
}
\description{
Preview color palette for Hurwitz Lab colors
}
