% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{get_hurwitz_colors}
\alias{get_hurwitz_colors}
\title{Extract colors as hex codes}
\usage{
get_hurwitz_colors(...)
}
\arguments{
\item{...}{Character names of hurwitz_colors}
}
\description{
Extract colors as hex codes
}
