% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{get_hurwitz_colours}
\alias{get_hurwitz_colours}
\title{Extract colors as hex codes}
\usage{
get_hurwitz_colours(...)
}
\arguments{
\item{...}{Character names of hurwitz_colours}
}
\description{
Extract colors as hex codes
}
