% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{get_hurwitz_pal}
\alias{get_hurwitz_pal}
\title{Return function to interpolate a color palette}
\usage{
get_hurwitz_pal(palette = "all", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in hurwitz_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a color palette
}
